Add WAF functions
This is a big diff, but the WAF is an important component of the
Fastly platform.

This adds the WAF functions to VCL:

Transformation functions:

- waf.cmdline, waf.cssDecode, waf.jsDecode, waf.htmlEntityDecode
- waf.hexEncode, waf.utf8ToHex
- waf.normalizePath, waf.normalizePathWin
- waf.urlDecodeUni, waf.urlDecodeUni_removeNulls

They need a bit of manual reimplementations in order to stricly match
what Fastly is doing.

Validation functions:
- waf.validateByteRange, waf.validateUrlEncoding, waf.validateUtf8Encoding

And the detection functions: waf.detectSQLi and waf.detectXSS.

Detection relies on libinjection. Maybe we could reimplement something
instead. In any case, that wouldn't match what's running on Fastly since
this is closed-source.

But libinjection is already very useful to test against common patterns.

Nits